+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove core-js, .travis.yml, azure-pipelines.yml
 * Add CI/CD and CodeQL
 * Add Rendering API support
+* Fix less-loader options, fix running tests, migrate from puppeteer to playwright-chromium
 
 ## 2.0.3
 * Fixed ticks formatting of fractional numbers